Produce actual and expected plan files for minidumps on ctest failures.
Currently when we run ctest, we don't see the failure whenever a test fails.
We have to run individual tests using `server/gporca_test -U testname` which
dumps the output to the screen and then manually cut paste the plan changes
into an editor to see the plan diffs.

This commit writes the actual and expected plans to files which we can diff
directly.

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>