Fix stats of equi join when join condtion is of type f(a) = b
Revert commit 9769f86237229fb5f173c867c07b644bde884ddc.
Above revert treats f(a) = b as a = b. This was an underestimation and hence was
earlier reverted.  But reverting this meant treating f(a) = b as unsupported
comparison and falling back to default stats, which estimates the cardinality
of this join as 40%. This was an overestimation.

Now for equi-join cases, when join condtion is of type f(a) = b, we follow the
formula suggested by Selinger et al.  The NDV of this join is equal to NDV(b).