Bump down the value of Epsilon
0.001 was too high a value for Epsilon. In billion+ rows table with highly
selective predicate, we ended up severly under estimating the cardinality of
predicate.

The Relcache Translator while calculating the MCV freq uses the comparison
`CStatistics::Epsilon < mcv_freq`. If this comparison fails, Relcache
translator will translate MCV freq as 0.  Before this fix, it was expecting the
MCV freq to be > 0.001. Which is too high a number for larger tables.  e.g. in
10 billion rows+ table, an MCV freq of 0.0001 will be passed as 0. But such a
selective predicate is actually selecting 1000000 rows. ORCA will though report
it as 1 row. Such cardinality misestimation can lead to picking an NLJ over
hashjoin or not picking an Index Scan.

Co-authored-by: Sambitesh Dash sdash@pivotal.io
Co-authored-by: Chris Hajas chajas@pivotal.io