Collapse inner and left outer joins into NAry joins
In the preprocessor, we only collapsed inner joins into NAry joins
so far. This commit also adds LOJs. The change is guarded by a
traceflag that is off by default, so this is dead code at this stage
until we deliver the needed optimizer changes.

We also change the algorithm to be more efficient.

- renamed PexprCollapseInnerJoins to PexprCollapseJoins
- added a list of integers to LogicalNAryJoin. These integers
  are present only when the NAry join contains at least one
  LOJ. The indexes help in associating join predicates to
  children.

Co-authored-by: Sambitesh Dash <sdash@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>