Add dependency info for LOJ edges
Added a new array to CJoinOrderDPv2 that tracks the needed tables
on the left-hand side of the LOJ. This can be used to determine whether
two components form a valid LOJ.

Also changed references in the code from LOJ to NIJ (non-inner join),
in anticipation of including semijoins into the DPv2 xform.

Also using the term "non-join vertex" for the level 1 components, that
term was already used.

Finally, changed the SEdge struct in CJoinOrder. This had a boolean
for whether the edge was for an LOJ. Replaced that with a number, so
we can associate an edge to the correct LOJ (NIJ) in DPv2.