Work in progress, new way to compute cost and limit # of expressions
Some renaming:

- non-join vertices => atoms   (easier name)
- component => group           (what we call "component" is not what
                                a component is in graph theory)

New data structures to describe join levels:

We still have an array of levels. Each level is now an array of SGroupInfo
structs. Each SGroupInfo struct has a bitset of atoms (a component) and
an expression. We also now have pointers to child SGroupInfos, to help
computing the cost.

Cost now has the cardinality of the join itself. To avoid computing stats
too many times, the stats are stored in each group and we have a new map to
look up already computed stats, given a bitset.

We no longer make a list of CExpressionArrays and then pick the cheapest
of each ExpressionArray. Instead, we delete expressions that aren't the
cheapest in their bitset immediately. To do that, we store all the
good expressions in the CJoinOrderDPv2, not in temporary variables.

The top k limit is now on the number of groups (aka bitsets aka components),
not the number of expressions, since we now only keep one expression
per group.

The top k limit is no longer a square function, it is now based on the
possible number of groups of a 10-way join (10 choose k). Right now we
don't pick the k cheapest groups, we just cut off at some point - needs
to be improved.

We only return one expression at the top level, needs to be changed.