Introduce a new scale factor algorithm for joins
This commit introduces a new algorithm for calculating the cumulative
scale factor for joins when estimating cardinality. The new algorithm
moderately decreases the impact of subsequent predicates by a sqrt
factor but only damps the predicate if there are multiple predicates on
the same two tables. Otherwise, we assume that there is no correlation
between predicates and thus don't need to be damped.

For example, if given the following join predicates:
    t1.a = t2.a AND t1.b = t2.b AND t2.b = t3.a

The algorithm would do the following:
    cumulative scale factor = (SF of t1.a = t2.a) * sqrt(SF of t1.b =
t2.b) * (SF of t2.b = t3.a)

In order for this to work, we need to keep track of the table from which
the predicate came. Thus, we pass along the mdid of the table through
the colref.

The algorithm is not the default and can be turned on using
`set optimizer_damping_factor_join = 0`.

Co-authored-by: Ashuka Xue <axue@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>