Implemented Query, Greedy, MinCard with the new DPv2 xform
The existing query, greedy and mincard xforms didn't handle the new NAry joins that
contained LOJs. To solve this, we integrated query, greedy and mincard into the DPv2
xform, using properties. We hope to re-use this property infrastructure in the future
if/when we improve the cost model used for DPv2.

Until now, we stored only the best join expression per group. With this commit, we
store the best expression for each unique property. Properties right now are the
type of join enumeration used, i. e. query, greedy, mincard and DPv2. So, we might
store a separate greedy, mincard and DPv2 expression, for example.

For a picture of some new data structures introduced, see the "Data structures for DPv2 join enumeration"
comment in file CJoinOrderDPv2.h.

Co-authored-by: Sambitesh Dash <sdash@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>