Check if there is direct dispatchable fileter
Consider two queries :

a) Explain Select RANK FROM ( select RANK() OVER ( order BY dist_key asc ) as
RANK, *  from foo )A WHERE '2014-06-29' = dist_key ;

b) Explain  Select RANK FROM ( select RANK() OVER ( order BY dist_key asc ) as
RANK, *  from foo WHERE '2014-06-29' = dist_key )A;

Query 'a' gives a plan like below with a filter over Window. So in spite of
having a filter on distribution key, we shouldn't direct dispatch.

	Filter: dist_key= '2014-06-29'::date
   	->  Window
         	Order By: dist_key

Query 'b' on other hand has a filter directly over a table scan, so we should
direct dispatch this plan.

We had a very simple heuristic to decide if a plan is direct dispatchable or
not. Now we add a check that if anywhere in the plan we run into a
'CPhyscialFilter' which is not above a scan, we bail out from direct
dispatching. We are erring on the side of caution here.